+++ CHANGELOG.md
@@ -11,6 +11,16 @@ all APIs might be changed.
 
 ## Unreleased - xxxx-xx-xx
 
+### New Features
+
+- The `QueryFragment` derive now supports renaming fields.
+
+### Bug Fixes
+
+- Underscore field names are now supported in schemas and for querying.
+- Field names with leading underscores will no longer have those leading
+  underscores removed.
+
 ## v0.13.2 - 2021-05-16
 
 This release is only of the `cynic` crate - other crates remain at 0.13.1